+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/base/models/DatabaseConfig.kt
@@ -1,5 +1,6 @@
 package com.github.insanusmokrassar.AutoPostTelegramBot.base.models
 
+import kotlinx.serialization.Optional
 import kotlinx.serialization.Serializable
 import org.jetbrains.exposed.sql.Database
 
@@ -8,8 +9,16 @@ data class DatabaseConfig(
     val url: String,
     val driver: String,
     val username: String,
-    val password: String
+    val password: String,
+    @Optional
+    val initAutomatically: Boolean = true
 ) {
+    init {
+        if (initAutomatically) {
+            connect()
+        }
+    }
+
     fun connect(): Database {
         return Database.connect(
             url,